requirements.txt
greenlet==3.3.0
numpy==2.2.6
pandas==2.3.3
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
